Extracted url builder logic to an utility script
